Analysis and comparison of the outcome of different models on the Bristol Airbnb listing for price prediction.
Analysis and comparison of the outcome of different models on the Bristol Airbnb listing for price prediction.
In this analysis, a complete end-to-end machine learning project is implemented using the Bristol Airbnb Listings dataset. The aim of the analysis is to provide a comparison of the results of different regression models that can be used to do predictions on the prices of the listings.

Many sources depict machine learning to be more of a trial and error approach rather than a one-shot single solution method. The following project analyses this hypothesis by applying the different models of machine learning on the sample case of the Bristol Airbnb listing.